requirements/prod.txt
alembic==1.14.1
asyncpg
psycopg-binary==0.30.0
bcrypt==4.2.1
email_validator==2.2.0
fastapi==0.115.8
passlib==1.7.4
pydantic-settings==2.7.1
PyJWT==2.10.1
python-multipart==0.0.20
SQLAlchemy==2.0.37
uvicorn==0.34.0
gpxpy==1.6.2
simplekml==1.3.6